+++ src/aaudio.rs
@@ -3,8 +3,10 @@
 #![cfg(target_os = "android")]
 
 use crate::{AudioOutputDevice, BaseAudioOutputDevice, OutputDeviceParameters};
-use aaudio::{AAudioStream, AAudioStreamBuilder, CallbackResult, Direction, Format};
-use std::{collections::VecDeque, error::Error};
+use aaudio::{
+    AAudioStream, AAudioStreamBuilder, CallbackResult, Direction, Format, PerformanceMode,
+};
+use std::error::Error;
 
 pub struct AAudioOutputDevice {
     stream: AAudioStream,
@@ -24,55 +26,32 @@ impl AudioOutputDevice for AAudioOutputDevice {
         C: FnMut(&mut [f32]) + Send + 'static,
         Self: Sized,
     {
+        let frame_count = params.channel_sample_count as i32;
         let mut stream = AAudioStreamBuilder::new()
             .map_err(convert_err)?
-            .set_buffer_capacity_in_frames(params.channel_sample_count as i32)
+            // Ensure double buffering is possible.
+            .set_buffer_capacity_in_frames(2 * frame_count)
             .set_channel_count(params.channels_count as i32)
             .set_format(Format::F32)
+            .set_sample_rate(params.sample_rate as i32)
             .set_direction(Direction::Output)
+            .set_performance_mode(PerformanceMode::LowLatency)
+            // Force the AAudio to give the buffer of fixed size.
+            .set_frames_per_data_callback(frame_count)
             .set_callbacks(
-                {
-                    let mut samples_queue = VecDeque::<f32>::new();
-                    move |_, data, num_samples_per_channel| {
-                        // Render another portion of data if needed.
-                        while num_samples_per_channel as usize
-                            > samples_queue.len() / params.channels_count
-                        {
-                            let write_pos = samples_queue.len();
+                move |_, data, num_frames| {
+                    let output_data = unsafe {
+                        std::slice::from_raw_parts_mut::<f32>(
+                            data.as_mut_ptr() as *mut f32,
+                            num_frames as usize * params.channels_count,
+                        )
+                    };
 
-                            let new_len = samples_queue.len()
-                                + params.channel_sample_count * params.channels_count;
+                    data_callback(output_data);
 
-                            while samples_queue.len() < new_len {
-                                samples_queue.push_back(0.0);
-                            }
-
-                            samples_queue.make_contiguous();
-
-                            let (left, _) = samples_queue.as_mut_slices();
-                            let dest_data = &mut left[write_pos..];
-                            debug_assert_eq!(
-                                dest_data.len(),
-                                params.channel_sample_count * params.channels_count
-                            );
-                            data_callback(dest_data);
-                        }
-
-                        let output_data = unsafe {
-                            std::slice::from_raw_parts_mut::<f32>(
-                                data.as_mut_ptr() as *mut f32,
-                                num_samples_per_channel as usize * params.channels_count,
-                            )
-                        };
-
-                        for sample in output_data {
-                            *sample = samples_queue.pop_back().expect("Queue underflow!")
-                        }
-
-                        CallbackResult::Continue
-                    }
+                    CallbackResult::Continue
                 },
-                |_, _| {},
+                |_, error| eprintln!("AAudio: an error has occurred - {:?}", error),
             )
             .open_stream()
             .map_err(convert_err)?;